+++ Changelog.md
@@ -1,3 +1,7 @@
+# Unreleased / 2014-07-123
+
+* `preprocesser` option (mainly to use [`css-whitespace`](https://github.com/reworkcss/css-whitespace))
+
 # 1.0.0 / 2014-07-14
 
 Merge with [rework-inline](https://www.npmjs.org/package/rework-inline) bring the following: